fix(trusty-memory): resolve OpenRouter key via shared function, split user_config module
Addresses code-critic re-review HIGH on #2977: dream_config_from_user_config
checked only config.toml's openrouter_api_key, while
build_consolidator_from_config also fell back to the OPENROUTER_API_KEY
process env var before choosing a backend. When config.toml had no key but
the daemon's env did, the helper picked the local-model id while the
consolidator built the OpenRouter backend, silently sending an Ollama tag
to OpenRouter every cycle — a regression vs pre-fix behavior for that
config shape.

- Add trusty_common::memory_core::semantic_consolidation::resolve_openrouter_api_key
  as the single source of truth for "configured value, else env fallback";
  build_consolidator_from_config and dream_config_from_user_config both call
  it now instead of each resolving the key independently.
- New tests: resolve_openrouter_api_key_{prefers_configured_value,falls_back_to_env}
  (trusty-common) and dream_config_from_user_config_prefers_openrouter_model_with_env_key
  (trusty-memory), all using a #[serial] + EnvVarGuard pattern to avoid
  racing other tests on the real env var.
- Split dream_config_from_user_config/load_user_config/LoadedUserConfig out
  of service/helpers.rs into a new service/user_config.rs module (the env-key
  fix pushed helpers.rs to 503 SLOC, over the 500 production cap); re-exported
  from service::mod unchanged so crate::service::X call sites are unaffected.

Closes #2593
